DayRow refactor
